refactor: member 패키지 변경
member, memberType을 한 패키지로 묶어 관리합니다.

Related to: #27
