+++ backend/src/main/java/com/prograngers/backend/entity/member/MemberType.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.entity;
+package com.prograngers.backend.entity.member;
 
 public enum MemberType {
     BASIC, KAKAO, GOOGLE, NAVER;
